requirements.txt
paddlepaddle-gpu
paddleocr
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.1
opencv-python
numpy
ultralytics
scipy
scikit-image
matplotlib
pillow
streamlit